+++ test/test_query.cpp
@@ -8377,8 +8377,9 @@ TEST(Query_MaximumSumAverage)
 
 TEST(Query_ReferDeletedLinkView)
 {
-    // This would segfault because the query refers a LinkView that had been removed by move_last_over
-    // It will now throw an exception instead.
+    // Queries and TableViews that depend on a deleted LinkList will now produce valid empty-like results
+    // (find() returns npos, find_all() returns empty TableView, sum() returns 0, etc.).
+    // They will no longer throw exceptions or crash.
     Group group;
     TableRef table = group.add_table("table");
     table->add_column_link(type_LinkList, "children", *table);
@@ -8389,14 +8390,30 @@ TEST(Query_ReferDeletedLinkView)
     Query q = table->where(links);
     TableView tv = q.find_all();
     
-    table->move_last_over(0);
+    // TableView that depends on LinkView soon to be deleted
+    TableView tv_sorted = links->get_sorted_view(1);
 
+    // Delete LinkList so LinkView gets detached
+    table->move_last_over(0);
+    CHECK(!links->is_attached());
+
+    // See if "Query that depends on LinkView" returns sane "empty"-like values
     CHECK_EQUAL(q.find_all().size(), 0);
+    CHECK_EQUAL(q.find(), npos);
     CHECK_EQUAL(q.sum_int(1), 0);
     CHECK_EQUAL(q.count(), 0);
     size_t rows;
     q.average_int(1, &rows);
     CHECK_EQUAL(rows, 0);
+
+    // See if "TableView that depends on LinkView" returns sane "empty"-like values
+    tv_sorted.average_int(1, &rows);
+    CHECK_EQUAL(rows, 0);
+
+    // Now check a "Query that depends on (TableView that depends on LinkView)"
+    Query q2 = table->where(&tv_sorted);
+    CHECK_EQUAL(q2.count(), 0);
+    CHECK_EQUAL(q2.find(), npos);
 
     CHECK(!links->is_attached());
     tv.sync_if_needed();